Split font and text file
